Refactor : moved package PostDto
